+++ lib/legion/extensions/slack.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/slack/version'
+require 'legion/extensions/slack/helpers/client'
+require 'legion/extensions/slack/runners/chat'
+require 'legion/extensions/slack/runners/user'
+require 'legion/extensions/slack/client'
 
 module Legion
   module Extensions